Rename `download_fossa_cli` to be terser
