+++ src/main.rs
@@ -9,8 +9,8 @@ use atty::Stream;
 use broker::api::remote::RemoteProvider;
 use broker::db;
 use broker::doc::crate_version;
-use broker::download_fossa_cli;
 use broker::ext::error_stack::IntoContext;
+use broker::fossa_cli;
 use broker::{config, ext::error_stack::ErrorHelper};
 use broker::{
     doc,
@@ -171,7 +171,7 @@ async fn main_run(args: config::RawBaseArgs) -> Result<(), Error> {
         .await
         .change_context(Error::InternalSetup)?;
 
-    let fossa_path = download_fossa_cli::ensure_fossa_cli(args.context())
+    let fossa_path = fossa_cli::ensure_fossa_cli(args.context())
         .await
         .change_context(Error::InternalSetup)?;
     debug!("fossa path: {:?}", fossa_path);
